gateway/requirements.txt
azure-core==1.29.5
azure-identity==1.14.1
azure-storage-blob==12.18.3
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.4
dnspython==2.4.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.24.1
msal-extensions==1.0.0
pika==1.3.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pymongo==4.5.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
